+++ packages/iov-bcp/types/connection.d.ts
@@ -122,8 +122,10 @@ export interface BcpQueryTag {
 }
 export interface BcpTxQuery {
     readonly id?: TransactionId;
-    /** send transaction to or from this address */
+    /** any send transaction to or from this address */
     readonly sentFromOrTo?: Address;
+    /** any transaction signed by this address */
+    readonly signedBy?: Address;
     /** chain-specific key value pairs that encode a query */
     readonly tags?: ReadonlyArray<BcpQueryTag>;
     readonly height?: number;