+++ packages/iov-bns/src/util.ts
@@ -137,8 +137,16 @@ function sentFromOrToTag(addr: Address): string {
   return `${key}='${value}'`;
 }
 
+function signedByTag(addr: Address): string {
+  const id = Uint8Array.from([...Encoding.toAscii("sigs:"), ...decodeBnsAddress(addr).data]);
+  const key = Encoding.toHex(id).toUpperCase();
+  const value = "s"; // "s" for "set"
+  return `${key}='${value}'`;
+}
+
 export function buildTxQuery(query: BcpTxQuery): QueryString {
   const sentComponents = query.sentFromOrTo !== undefined ? [sentFromOrToTag(query.sentFromOrTo)] : [];
+  const signedByComponents = query.signedBy !== undefined ? [signedByTag(query.signedBy)] : [];
   const tagComponents = query.tags !== undefined ? query.tags.map(tag => `${tag.key}='${tag.value}'`) : [];
   // In Tendermint, hash can be lower case for search queries but must be upper case for subscribe queries
   const hashComponents = query.id !== undefined ? [`tx.hash='${query.id}'`] : [];
@@ -148,6 +156,7 @@ export function buildTxQuery(query: BcpTxQuery): QueryString {
 
   const components: ReadonlyArray<string> = [
     ...sentComponents,
+    ...signedByComponents,
     ...tagComponents,
     ...hashComponents,
     ...heightComponents,