+++ packages/iov-bns/types/index.d.ts
@@ -1,5 +1,5 @@
 export { bnsCodec } from "./bnscodec";
 export { bnsConnector } from "./bnsconnector";
 export { BnsConnection } from "./bnsconnection";
-export { bnsNonceTag, bnsSwapQueryTag } from "./tags";
+export { bnsSwapQueryTag } from "./tags";
 export { ChainAddressPair, BnsBlockchainNft, BnsBlockchainsByChainIdQuery, BnsBlockchainsQuery, BnsUsernamesByChainAndAddressQuery, BnsUsernamesByOwnerAddressQuery, BnsUsernamesByUsernameQuery, BnsUsernamesQuery, BnsUsernameNft, BnsTx, AddAddressToUsernameTx, RegisterBlockchainTx, RegisterUsernameTx, RemoveAddressFromUsernameTx, } from "./types";