+++ packages/iov-rise/src/riseconnection.ts
@@ -304,8 +304,8 @@ export class RiseConnection implements BcpConnection {
   }
 
   public async searchTx(query: BcpTxQuery): Promise<ReadonlyArray<ConfirmedTransaction>> {
-    if (query.height || query.tags) {
-      throw new Error("Query by height and tags not supported");
+    if (query.height || query.tags || query.signedBy) {
+      throw new Error("Query by height, tags or signedBy not supported");
     }
 
     if (query.id !== undefined) {
@@ -334,8 +334,8 @@ export class RiseConnection implements BcpConnection {
   }
 
   public liveTx(query: BcpTxQuery): Stream<ConfirmedTransaction | FailedTransaction> {
-    if (query.height || query.tags) {
-      throw new Error("Query by height or tags not supported");
+    if (query.height || query.tags || query.signedBy) {
+      throw new Error("Query by height, tags or signedBy not supported");
     }
 
     if (query.id !== undefined) {